+++ Koronavilkku/KoronavilkkuTests/Localization/LocalizationTests.swift
@@ -2,7 +2,7 @@ import XCTest
 @testable import Koronavilkku
 
 class LocalizationTests: XCTestCase {
-    private let supportedLanguages = ["fi"]
+    private let supportedLanguages = ["fi", "sv", "en"]
     
     func testLocalizationsDone() {
         verifyTranslations(from: Translation.self)
@@ -18,6 +18,7 @@ class LocalizationTests: XCTestCase {
         verifyTranslations(from: SymptomsViewController.Text.self)
         verifyTranslations(from: SettingsViewController.Text.self)
         verifyTranslations(from: SymptomsElement.Text.self)
+        verifyTranslations(from: SelectMunicipalityViewController.Text.self)
     }
     
     private func verifyTranslations<T: Localizable & CaseIterable>(from namespace: T.Type) {