+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Schedule the EN background task to run max 6 times a day to prevent it being run excessively by the system
 - Suppress any errors encountered while updating the municipality list in the background task
 - Ensure background task being rescheduled even if current task is unexpectedly terminated 
+- Localisation updates
 
 ### Added
 - REST API tests
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Proper TEK generation in MockExposureManager
 - Git commit hash to version information in Settings view
 - Show the change log in Xcode workspace
+- English translation
 
 ## 1.0.1